pyproject.toml
[tool.poetry]
name = "smartmart-backend"
version = "0.1.0"
description = ""
authors = ["cloco-sagar",]
package-mode=false

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[tool.ruff]
lint.select = ["E", "F", "W", "C"]
lint.ignore = ["E501"]  # Ignore long-line errors since Black handles it
target-version = "py312"

[tool.ruff.lint.isort]
case-sensitive = true
force-single-line = true
section-order = ["future", "standard-library", "first-party", "third-party", "local-folder"]


[tool.black]
line-length = 79
target-version = ['py312']
include = '\.pyi?$'


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
ruff = "^0.9.3"
alembic = "^1.14.1"
pydantic-settings = "^2.7.1"
psycopg2-binary = "^2.9.10"
click = "^8.1.8"
bcrypt = "^4.2.1"
pyjwt = "^2.10.1"
pytz = "^2024.2"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
